fix naval exploration for good fix army recruitment bug AI doesn't build wonders when it doesn't stand a chance

+++ CvGameCoreDLL_Expansion2/CvMinorCivAI.cpp
@@ -963,7 +963,7 @@ void CvMinorCivAI::RecalculateRewards(PlayerTypes ePlayer)
 	if (!GetPlayer()->isAlive())
 		return;
 
-	if (m_QuestsGiven.empty() || m_QuestsGiven.size() <= ePlayer)
+	if (m_QuestsGiven.empty() || m_QuestsGiven.size() <= (size_t)ePlayer)
 		return;
 
 	QuestListForPlayer::iterator itr_quest;
